+++ test/fixtures/generator-and-minimizer-resize-query.js
@@ -1,0 +1,15 @@
+require("./loader-test.png?width=100");
+require("./loader-test.png?w=150");
+require("./loader-test.png?height=200");
+require("./loader-test.png?h=250");
+require("./loader-test.png?width=300&height=auto");
+require("./loader-test.png?width=auto&height=320");
+require("./loader-test.png?width=350&height=350");
+
+require("./loader-test.png?as=webp&width=100");
+require("./loader-test.png?as=webp&w=150");
+require("./loader-test.png?as=webp&height=200");
+require("./loader-test.png?as=webp&h=250");
+require("./loader-test.png?as=webp&width=300&height=auto");
+require("./loader-test.png?as=webp&width=auto&height=320");
+require("./loader-test.png?as=webp&width=350&height=350");