+++ types/index.d.ts
@@ -54,11 +54,17 @@ export = ImageMinimizerPlugin;
  * @typedef {InferDefaultType<T> | undefined} BasicTransformerOptions
  */
 /**
+ * @typedef {Object} ResizeOptions
+ * @property {number} [width]
+ * @property {number} [height]
+ * @property {boolean} [enabled]
+ */
+/**
  * @template T
  * @callback BasicTransformerImplementation
  * @param {WorkerResult} original
  * @param {BasicTransformerOptions<T>} [options]
- * @returns {Promise<WorkerResult>}
+ * @returns {Promise<WorkerResult | null>}
  */
 /**
  * @typedef {Object} BasicTransformerHelpers
@@ -187,6 +193,7 @@ declare namespace ImageMinimizerPlugin {
     CustomOptions,
     InferDefaultType,
     BasicTransformerOptions,
+    ResizeOptions,
     BasicTransformerImplementation,
     BasicTransformerHelpers,
     TransformerFunction,
@@ -303,10 +310,15 @@ type CustomOptions = {
 };
 type InferDefaultType<T> = T extends infer U ? U : CustomOptions;
 type BasicTransformerOptions<T> = InferDefaultType<T> | undefined;
+type ResizeOptions = {
+  width?: number | undefined;
+  height?: number | undefined;
+  enabled?: boolean | undefined;
+};
 type BasicTransformerImplementation<T> = (
   original: WorkerResult,
   options?: BasicTransformerOptions<T>
-) => Promise<WorkerResult>;
+) => Promise<WorkerResult | null>;
 type BasicTransformerHelpers = {
   setup?: (() => void) | undefined;
   teardown?: (() => void) | undefined;